fix: use tx.mode instead of tx.type
