+++ .changeset/eighty-items-tickle.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-aleo": minor
+---
+
+fix: use tx.mode instead of tx.type in coin-aleo
