+++ libs/coin-modules/coin-aleo/src/bridge/createTransaction.test.ts
@@ -15,7 +15,7 @@ describe("createTransaction", () => {
       useAllAmount: false,
       recipient: "",
       fees: new BigNumber(0),
-      type: TRANSACTION_TYPE.TRANSFER_PUBLIC,
+      mode: TRANSACTION_TYPE.TRANSFER_PUBLIC,
     });
   });
 });
